improvement(docker): several improvements to Docker Compose DEV environment
Previously ihe EWP Node was compiled into a single container for the DEV environment not allowing hot reloading of the frontend. It was splitted into two separate containers, one for frontend and other for backend.
Besides, the files necessary for the EWP Registry container are now created dynamically when building the Docker image.